Update to new LoRaWAN APIs. Fix session reset bug
seperate device behavior logic

clean up "use"

do a small refactor
