+++ src/udp_radio.rs
@@ -1,14 +1,8 @@
 #![macro_use]
-use super::{
-    config, debugln, prometheus_service as prometheus, prometheus_service::Stat, udp_runtime,
-    INSTANT,
-};
+use super::{config, debugln, udp_runtime, INSTANT};
 use heapless::consts::*;
 use heapless::Vec as HVec;
-use lorawan_device::{
-    self as lorawan, radio, Device as LorawanDevice, Event as LorawanEvent,
-    Response as LorawanResponse, Timings,
-};
+use lorawan_device::{radio, Timings};
 use semtech_udp::{push_data, push_data::RxPk, Down, Packet, StringOrNum};
 use std::time::Duration;
 use tokio::sync::{
@@ -81,225 +75,6 @@ pub struct UdpRadioRuntime {
     time: Instant,
 }
 
-pub fn pretty_device(creds: &lorawan::Credentials) -> String {
-    let mut bytes: Vec<u8> = Vec::new();
-    bytes.extend(creds.deveui());
-    bytes.reverse();
-    let hex = hex::encode(&bytes);
-    hex.to_uppercase()[12..].to_string()
-}
-
-pub async fn need_to_schedule_packet(
-    lorawan_sender: &mut Sender<IntermediateEvent>,
-    lorawan: &mut LorawanDevice<UdpRadio>,
-    fcnt_before_rejoin: Option<usize>,
-) -> bool {
-    if let (Some(fcnt_threshold), Some(fcnt)) = (fcnt_before_rejoin, lorawan.get_fcnt_up()) {
-        if fcnt >= fcnt_threshold as u32 {
-            lorawan_sender
-                .send(IntermediateEvent::NewSession)
-                .await
-                .unwrap();
-            return false;
-        }
-    }
-    true
-}
-
-pub async fn schedule_packet(
-    lorawan_sender: &mut Sender<IntermediateEvent>,
-    lorawan: &mut LorawanDevice<UdpRadio>,
-    transmit_delay: u64,
-) {
-    let delay = if lorawan.get_radio().jitter {
-        (super::get_random_u32() & 0x7F) as u64
-    } else {
-        0
-    };
-
-    let mut sender = lorawan_sender.clone();
-    tokio::spawn(async move {
-        delay_for(Duration::from_millis(transmit_delay as u64 + delay)).await;
-        sender.send(IntermediateEvent::SendPacket).await.unwrap();
-    });
-}
-
-pub async fn run_loop(
-    mut lorawan_receiver: Receiver<IntermediateEvent>,
-    mut lorawan_sender: Sender<IntermediateEvent>,
-    mut lorawan: LorawanDevice<UdpRadio>,
-    mut prometheus: Option<Sender<prometheus::Message>>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    lorawan_sender
-        .try_send(IntermediateEvent::NewSession)
-        .unwrap();
-
-    loop {
-        let device_ref = pretty_device(lorawan.get_credentials());
-        let transmit_delay = lorawan.get_radio().config.transmit_delay();
-        let fcnt_before_rejoin = lorawan.get_radio().config.fcnt_before_rejoin();
-
-        if let Some(event) = lorawan_receiver.recv().await {
-            let mut time = None;
-            let (new_state, response) = match event {
-                IntermediateEvent::NewSession => {
-                    // if jitter is enabled, we'll delay 0-127 ms
-                    let delay = if lorawan.get_radio().jitter {
-                        (super::get_random_u32() & 0x7F) as u64
-                    } else {
-                        0
-                    };
-
-                    delay_for(Duration::from_millis(1000 + delay as u64)).await;
-
-                    debugln!("{}: Creating Session", device_ref);
-                    let event = LorawanEvent::NewSession;
-                    lorawan.handle_event(event)
-                }
-                IntermediateEvent::SendPacket => {
-                    let data = [12, 3, 54, 54, 123, 23, 13, 14, 15, 16];
-                    let mut ret = lorawan.send(&data, 2, true);
-                    debugln!(
-                        "{}: Sending DataUp, FcntUp = {}",
-                        device_ref,
-                        ret.0.get_fcnt_up().unwrap() - 1
-                    );
-                    ret
-                }
-                IntermediateEvent::Rx(packet, time_received) => {
-                    time = Some(time_received);
-                    lorawan.handle_event(LorawanEvent::RadioEvent(radio::Event::PhyEvent(
-                        packet.into(),
-                    )))
-                }
-                IntermediateEvent::Timeout(id) => {
-                    if lorawan.get_radio().most_recent_timeout(id) {
-                        lorawan.handle_event(LorawanEvent::Timeout)
-                    } else {
-                        (lorawan, Ok(LorawanResponse::Idle))
-                    }
-                }
-            };
-
-            lorawan = new_state;
-
-            match response {
-                Ok(response) => match response {
-                    LorawanResponse::TimeoutRequest(delay) => {
-                        lorawan.get_radio().timer(delay).await;
-                    }
-                    LorawanResponse::NoJoinAccept => {
-                        debugln!("{}: No JoinAccept Received", device_ref);
-                        if let Some(ref mut sender) = prometheus {
-                            sender
-                                .send(prometheus::Message::Stat(device_ref, Stat::JoinTimeout))
-                                .await?
-                        }
-                        // if the Join Request failed try again
-                        lorawan_sender
-                            .send(IntermediateEvent::NewSession)
-                            .await
-                            .unwrap();
-                    }
-                    LorawanResponse::NewSession => {
-                        if let Some(t) = time {
-                            debugln!(
-                                "{}: JoinSuccess  [{} ms to spare] {:?}",
-                                device_ref,
-                                t,
-                                lorawan.get_session_keys().unwrap()
-                            );
-                            if let Some(ref mut sender) = prometheus {
-                                sender
-                                    .send(prometheus::Message::Stat(
-                                        device_ref,
-                                        Stat::JoinResponse(t),
-                                    ))
-                                    .await?
-                            }
-                        }
-                        schedule_packet(&mut lorawan_sender, &mut lorawan, transmit_delay).await;
-                    }
-                    LorawanResponse::Idle => (),
-                    LorawanResponse::NoAck => {
-                        debugln!("{}: NoAck", device_ref);
-                        if let Some(ref mut sender) = prometheus {
-                            sender
-                                .send(prometheus::Message::Stat(device_ref, Stat::DownlinkTimeout))
-                                .await?
-                        }
-
-                        let mut sender = lorawan_sender.clone();
-                        tokio::spawn(async move {
-                            delay_for(Duration::from_millis(transmit_delay as u64)).await;
-                            sender.send(IntermediateEvent::SendPacket).await.unwrap();
-                        });
-                    }
-                    LorawanResponse::ReadyToSend => {
-                        debugln!(
-                            "{}: No downlink received but none expected - ready to send again",
-                            device_ref
-                        );
-                        schedule_packet(&mut lorawan_sender, &mut lorawan, transmit_delay).await;
-                    }
-                    LorawanResponse::DataDown(fcnt_down) => {
-                        if let Some(t) = time {
-                            debugln!(
-                                "{}: DataDown [{} ms to spare], FcntDown = {} ",
-                                device_ref,
-                                t,
-                                fcnt_down
-                            );
-                            if let Some(ref mut sender) = prometheus {
-                                sender
-                                    .send(prometheus::Message::Stat(
-                                        device_ref,
-                                        Stat::DownlinkResponse(t),
-                                    ))
-                                    .await?
-                            }
-                        }
-
-                        if need_to_schedule_packet(
-                            &mut lorawan_sender,
-                            &mut lorawan,
-                            fcnt_before_rejoin,
-                        )
-                        .await
-                        {
-                            schedule_packet(&mut lorawan_sender, &mut lorawan, transmit_delay)
-                                .await;
-                        }
-                    }
-                    LorawanResponse::Rxing => {
-                        debugln!("{}: Receiving", device_ref);
-                    }
-                    _ => (),
-                },
-                Err(err) => match err {
-                    lorawan::Error::Radio(_) => (),
-                    lorawan::Error::Session(e) => {
-                        use lorawan::session::Error;
-                        match e {
-                            Error::RadioEventWhileIdle
-                            | Error::RadioEventWhileWaitingForRxWindow => (),
-                            _ => panic!("LoRaWAN Error Session {:?}\r\n", e),
-                        }
-                    }
-                    lorawan::Error::NoSession(e) => {
-                        use lorawan::no_session::Error;
-                        match e {
-                            Error::RadioEventWhileIdle
-                            | Error::RadioEventWhileWaitingForJoinWindow => (),
-                            _ => panic!("LoRaWAN Error NoSession {:?}\r\n", e),
-                        }
-                    }
-                },
-            }
-        }
-    }
-}
-
 impl UdpRadioRuntime {
     pub async fn run(&mut self) -> Result<(), Box<dyn std::error::Error>> {
         loop {
@@ -366,6 +141,14 @@ pub struct UdpRadio {
 }
 
 impl UdpRadio {
+    pub fn config(&self) -> &config::Device {
+        &self.config
+    }
+
+    pub fn jitter(&self) -> bool {
+        self.jitter
+    }
+
     pub fn new(
         sender: Sender<udp_runtime::TxMessage>,
         receiver: broadcast::Receiver<udp_runtime::RxMessage>,
@@ -428,7 +211,9 @@ impl UdpRadio {
     }
 }
 
+#[derive(Debug)]
 pub enum Error {}
+#[derive(Debug)]
 pub enum Response {}
 
 impl radio::PhyRxTx for UdpRadio {
@@ -503,10 +288,10 @@ impl radio::PhyRxTx for UdpRadio {
 }
 
 impl Timings for UdpRadio {
-    fn get_rx_window_offset_ms(&mut self) -> i32 {
+    fn get_rx_window_offset_ms(&self) -> i32 {
         20
     }
-    fn get_rx_window_duration_ms(&mut self) -> u32 {
+    fn get_rx_window_duration_ms(&self) -> u32 {
         100
     }
 }
